pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-usage-bar"
version = "0.1.0"
description = "macOS menu bar app showing real-time Claude usage from local session files"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "watchdog>=4.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
    # Platform-specific renderers — installed automatically by pip extras
    "rumps>=0.4.0; sys_platform == 'darwin'",
]

[project.optional-dependencies]
linux = [
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "keyring>=24.0.0",
]
windows = [
    "pystray>=0.19.0",
    "pillow>=10.0.0",
    "keyring>=24.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
claude-usage-bar = "claude_usage_bar.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_usage_bar"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
